Further work on static analysis utility.
- Renamed eonz.reflect.utils to eonz.reflect.analysis

- Created the concept of closures, allowing us to determine when a 
variable is captured by a closure.

- Added utilities to the Variable value reference class that enumerate 
its assignment references.

- Value:references([predicate]) now takes a predicate that will be used 
to construct a subset of the references registered to the Value 
instance.

- The analyzers now correctly apply the 'parameter' category to 
variables to distinguish them from declared locals. The category is used 
on function parameters, but is also used on iterator variables in for 
statements.

These changes pave the way towards the ability to statically reason 
about the contents of variables between statements where the variable is 
not liften by any closures.